Changing Article.Id type from int to long to prevent primary key overflow.

+++ NewsPicker.Database/Controllers/ArticlesController.cs
@@ -30,7 +30,7 @@ namespace NewsPicker.Database.Controllers
             _db.SaveChanges();
         }
 
-        public void UpdateEngagementCount(int articleId, int engagementCount)
+        public void UpdateEngagementCount(long articleId, int engagementCount)
         {
             var article = _db.Articles.Find(articleId);
